requirements.txt
discord.py==1.5.1
requests-futures==1.0.0
python-dotenv==0.15.0